requirements.txt
python-dotenv>=1.0.0
stripe>=10.0.0
fastapi
uvicorn 
pydantic
sqlalchemy
psycopg2-binary 
python-dotenv
stripe
